requirements.txt
python-dotenv
google-auth-oauthlib
google-api-python-client
spotipy